Export getInitials to utils package
