+++ packages/components/src/Avatar/README.md
@@ -1,29 +1,25 @@
 ### Simple avatar with name (initial only)
 
 ```jsx
-const Avatar = require("./Avatar").default
-;<Avatar name="Franklin Green" />
+<Avatar name="Franklin Green" />
 ```
 
 ### Show full name next to the avatar circle
 
 ```jsx
-const Avatar = require("./Avatar").default
-;<Avatar showName name="Franklin Green" />
+<Avatar showName name="Franklin Green" />
 ```
 
 ### Color the avatar circle with a custom or named color
 
 ```jsx
-const Avatar = require("./Avatar").default
-;<Avatar showName color="#142222" name="Franklin Green" />
+<Avatar showName color="#142222" name="Franklin Green" />
 ```
 
 ### Automatically assign colors that are deterministic by name
 
 ```jsx
-const Avatar = require("./Avatar").default
-;<>
+<>
   <Avatar showName name="Franklin Green" />
   <Avatar showName name="Tejas Kumar" />
   <Avatar showName name="Fabien Bernard" />
@@ -35,15 +31,13 @@ const Avatar = require("./Avatar").default
 This will automatically hide the initials
 
 ```jsx
-const Avatar = require("./Avatar").default
-;<Avatar photo="http://thecatapi.com/api/images/get?format=src&size=small" name="Franklin Green" />
+<Avatar photo="http://thecatapi.com/api/images/get?format=src&size=small" name="Franklin Green" />
 ```
 
 ### Should be also beautiful on a dark background
 
 ```jsx
-const Avatar = require("./Avatar").default
-;<div style={{ backgroundColor: "#333333", padding: 10 }}>
+<div style={{ backgroundColor: "#333333", padding: 10 }}>
   <Avatar photo="http://thecatapi.com/api/images/get?format=src&size=small" name="Franklin Green" />
 </div>
 ```
