+++ packages/utils/src/__tests__/text.test.ts
@@ -1,6 +1,6 @@
-import { getInitials } from "../Avatar/Avatar"
+import { getInitials } from "../text"
 
-describe("Avatar", () => {
+describe("text", () => {
   describe("getInitial", () => {
     ;[
       { value: "Tejas Kumar", expected: "TK" },
@@ -8,6 +8,7 @@ describe("Avatar", () => {
       { value: "Tejas", expected: "T" },
       { value: "Lucrèce Rolland Nevière", expected: "LN" },
       { value: "Jean-pierre Bernard", expected: "JB" },
+      { value: "", expected: "" },
     ].map(({ value, expected }) =>
       it(`should return ${expected} for ${value}`, () => expect(getInitials(value)).toEqual(expected)),
     )
